+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 - Zip the json files for lower disk usage.  For daily archives, keep one 
   zip file for a year worth of data.  
 
+## Release 2023-10-01
+- IESO work.  Archive the rt_zonal_demand and rt_generation.  
+
 ## Release 2023-09-27
 - Change Plotly.now arguments, require a file.  Add Plotly.exportJs.